qemu: move to go-qemu/qemu, update documentation
