+++ qemu/block.go
