+++ qemu/block_test.go
