+++ qemu/cpu.go
