+++ qemu/domain.go
@@ -16,7 +16,7 @@
 package qemu
 
 //go:generate stringer -type=Status -output=string.gen.go
-//go:generate ./scripts/prependlicense.sh string.gen.go
+//go:generate ../scripts/prependlicense.sh string.gen.go
 
 import (
 	"encoding/json"
