+++ qemu/domain_test.go
