+++ qemu/string.gen.go
