+++ qemu/tempfile.go
