+++ qemu/version_test.go
